backend/requirements.txt
# Job Eco System - Backend Dependencies
# Install: pip install -r requirements.txt

python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
pymongo==4.10.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
pydantic==2.10.3
email-validator==2.2.0
requests==2.32.3
